+++ python_modules/dagster/dagster/_core/definitions/external_asset.py
